+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from "vscode";
 import { escapeMarkdownCodeBlock, profile } from "./utils";
 import { config } from "./config";
 import { jsonStringParser } from "./providers/json";
+import { jsoncStringParser } from "./providers/jsonc";
 
 function registerStringHoverProvider(
   selector: vscode.DocumentSelector,
@@ -35,7 +36,8 @@ function registerStringHoverProvider(
 
 export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
-    registerStringHoverProvider("json", jsonStringParser)
+    registerStringHoverProvider("json", jsonStringParser),
+    registerStringHoverProvider("jsonc", jsoncStringParser)
   );
 }